utill::rpc restructured
