+++ crate/utils/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod api;
 pub mod custody;
 pub mod eip1559;
-pub mod registry;
+pub mod rpc;
