+++ crate/utils/src/rpc/mod.rs
@@ -1,5 +1,3 @@
-pub mod api;
 pub mod rpc;
 
-pub use api::*;
 pub use rpc::*;
